format

+++ src/main/java/com/capstone/ar_guideline/dtos/responses/Wallet/WalletResponse.java
@@ -17,7 +17,7 @@ public class WalletResponse {
 
   private String userId;
 
-  private BigDecimal balance;
+  private Long balance;
 
   private String currency;
 }
